Refactored more views to improve code quality/readability/organization. Also deleted LoginView since we don't use it in our app
