+++ HyprRun/Views/Sub-views/NestedHomeViews.swift
@@ -42,4 +42,29 @@ extension HomeView {
       alignment: .topLeading)
     .padding()
   }
+  
+  func homeRewindView() -> some View {
+    // TODO: Add an EmptyState view to show "no runs - go for a run!" or something
+    
+    return VStack {
+      ScrollView {
+        HStack {
+          let miles = round(self.runViewModel.runs.map({ $0.distance}).reduce(0, +) * 100)/100.0
+          let count = self.runViewModel.runs.count
+          let avgDist: Double = round((miles / (Double(count))) * 100)/100.0
+          VStack {
+            Text("\(miles)").font(.custom("HelveticaNeue-Bold", fixedSize: 28))
+            Text("mi. ran").font(.custom("HelveticaNeue", fixedSize: 16))
+          }
+          Spacer()
+            .frame(width: 40)
+          VStack(alignment: .leading, spacing: 4) {
+            Text("Runs completed: \(count)")
+            Text("Avg. distance: \(avgDist)")
+          }
+        }
+      }
+    }
+    .onAppear(perform: self.runViewModel.retrieveRuns)
+  }
 }
